Reorganize notification modules
