+++ pom.xml
@@ -169,6 +169,7 @@
     <module>cdap-standalone</module>
     <module>cdap-distributions</module>
     <module>cdap-archetypes</module>
+    <module>cdap-notifications-api</module>
   </modules>
 
   <dependencyManagement>
